收件列表精确查询与模糊查询

+++ src/main/java/com/parcelhub/utils/JwtUtil.java
@@ -106,7 +106,7 @@ public class JwtUtil {
                 .getBody();
     }
 
-    public static long getUserId(HttpServletRequest request) throws Exception {
+    public static int getUserId(HttpServletRequest request) throws Exception {
         try
         {
             String token = request.getHeader("token");
